refactor: Move user menu and notifications to `app/layout`
