+++ app/layout/components/AppToolbar.tsx
@@ -16,8 +16,9 @@ import {
 import * as React from "react";
 import { Link as NavLink } from "../../common/Link.js";
 import { useCurrentUser } from "../../core/auth.js";
-import { NotificationsMenu, UserMenu } from "../../menus/index.js";
+import { NotificationsMenu } from "./NotificationsMenu.js";
 import { ThemeButton } from "./ThemeButton.js";
+import { UserMenu } from "./UserMenu.js";
 
 export function AppToolbar(props: AppToolbarProps): JSX.Element {
   const { sx, ...other } = props;
