--- app/menus/index.ts
@@ -1,5 +1,0 @@
-/* SPDX-FileCopyrightText: 2014-present Kriasoft */
-/* SPDX-License-Identifier: MIT */
-
-export { NotificationsMenu } from "./NotificationsMenu.js";
-export { UserMenu } from "./UserMenu.js";
